\name{@ANNOBJPREFIX@.db}
\alias{@ANNOBJPREFIX@.db}
\alias{@ANNOBJPREFIX@}
\title{Bioconductor annotation data package}
\description{

  Welcome to the @ANNOBJPREFIX@.db annotation Package.  This is an
  organism specific package.  The purpose is to provide detailed
  information about the species abbreviated in the second part of the
  package name @ANNOBJPREFIX@.db.  "Mm" is for Mus musculus and "Rn" is for
  Rattus norvegicus.  This package is updated biannually.
  
  You can learn what objects this package supports with the following command:

  \code{ls("package:@ANNOBJPREFIX@.db")} 
  
  Each of these objects has their own manual page detailing where relevant
  data was obtained along with examples of how to use it.  Many of these
  objects also have a reverse map available.  When this is true, expect
  to usually find relevant information on the same manual page as the
  forward map.
}
\examples{
  ls("package:@ANNOBJPREFIX@.db")
}
\keyword{datasets}

